third_party/WebKit/LayoutTests/fast/canvas/canvas-2d-imageData-create-nonfinite-expected.txt

Test the argument bounds of canvas createImageData.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.createImageData(Infinity, Infinity) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createImageData(Infinity, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createImageData(-Infinity, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createImageData(10, Infinity) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createImageData(10, -Infinity) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createImageData(NaN, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createImageData(10, NaN) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS successfullyParsed is true

TEST COMPLETE
